# api/openapi.yaml
openapi: 3.0.3
info:
  title: glamor daemon API
  version: 0.1.0
  description: |
    HTTP API и WS-события демона glamord (D-02: спека — source of truth,
    Go strict-server + модели через oapi-codegen, TS-типы через
    openapi-typescript — см. уточнение D-02 от 2026-08-16).
    Все мутации идемпотентны (D-12): header `Idempotency-Key`.
    Auth: Bearer-токен (D-08) на всё, кроме /healthz; для /ws — query-параметр.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
        description: динамический порт демона (D-08), см. ~/.glamor/daemon.json
...
paths:
...
  /projects:
    get:
      tags: [projects]
      operationId: listProjects
      summary: Список проектов
      responses:
        "200":
          description: проекты
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [projects]
      operationId: createProject
      summary: Зарегистрировать проект
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "200":
          description: проект с таким path уже существует (идемпотентность)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /projects/{id}:
    delete:
      tags: [projects]
      operationId: deleteProject
      summary: Удалить проект КАСКАДНО (раны, стадии, события, гейты, заметки)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: удалён
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        default:
          $ref: "#/components/responses/Error"
    get:
      tags: [projects]
      operationId: getProject
      summary: Проект с активными ранами и счётчиком открытых гейтов
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: проект
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [projects]
      operationId: patchProject
      summary: Обновить ide_command / default_branch
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchProjectRequest"
      responses:
        "200":
          description: обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /projects/{id}/pipelines:
    get:
      tags: [pipelines]
      operationId: listProjectPipelines
      summary: Пайплайны проекта (+ глобальные)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: пайплайны
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pipeline"
        default:
          $ref: "#/components/responses/Error"
  /pipelines/{id}:
    get:
      tags: [pipelines]
      operationId: getPipeline
      summary: Версия пайплайна со списком всех версий
      parameters:
        - $ref: "#/components/parameters/PipelineId"
      responses:
        "200":
          description: пайплайн
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineDetail"
        default:
          $ref: "#/components/responses/Error"
  /pipelines:
    post:
      tags: [pipelines]
      operationId: createPipeline
      summary: Создать пайплайн (первая версия)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePipelineRequest"
      responses:
        "201":
          description: создан (версия 1)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipeline"
        default:
          $ref: "#/components/responses/Error"
  /pipelines/import:
    post:
      tags: [pipelines]
      operationId: importPipeline
      summary: Импорт пайплайна из YAML (T-21)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPipelineRequest"
      responses:
        "201":
          description: импортирован (новый пайплайн или новая версия)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipeline"
        default:
          $ref: "#/components/responses/Error"
  /pipelines/{id}/versions:
    get:
      tags: [pipelines]
      operationId: listPipelineVersions
      summary: Все версии пайплайна
      parameters:
        - $ref: "#/components/parameters/PipelineId"
      responses:
        "200":
          description: версии (desc)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pipeline"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [pipelines]
      operationId: createPipelineVersion
      summary: Новая версия пайплайна (правка = новая версия, T-21)
      parameters:
        - $ref: "#/components/parameters/PipelineId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePipelineVersionRequest"
      responses:
        "201":
          description: создана версия N+1 с parent_version_id=id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipeline"
        default:
          $ref: "#/components/responses/Error"
  /pipeline-versions/{vid}:
    get:
      tags: [pipelines]
      operationId: getPipelineVersion
      summary: Конкретная версия пайплайна
      parameters:
        - name: vid
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: версия
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pipeline"
        default:
          $ref: "#/components/responses/Error"
  /pipeline-versions/{vid}/export:
    get:
      tags: [pipelines]
      operationId: exportPipelineVersion
      summary: Экспорт версии в YAML (T-21)
      parameters:
        - name: vid
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: YAML пайплайна (стабильный порядок ключей)
          content:
            text/yaml:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
  /runs:
    post:
      tags: [runs]
      operationId: createRun
      summary: Создать ран (идемпотентно, D-12)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "200":
          description: повтор Idempotency-Key — возвращён первый ран
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
    get:
      tags: [runs]
      operationId: listRuns
      summary: Список ранов с фильтрами
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
            format: int64
        - name: pipeline_id
          in: query
          schema:
            type: integer
            format: int64
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/RunState"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: раны
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
...
  /runs/{id}/stop:
    post:
      tags: [runs]
      operationId: stopRun
      summary: Остановить ран (идемпотентно, D-14)
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: остановлен (или уже остановлен)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
  /runs/{id}/resume:
    post:
      tags: [runs]
      operationId: resumeRun
      summary: Ручной рестарт failed/interrupted рана
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: ран возобновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        default:
          $ref: "#/components/responses/Error"
  /runs/{id}/notes:
    post:
      tags: [runs]
      operationId: createNote
      summary: Queue note к ближайшему событию рана (D-22)
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: заметка поставлена в очередь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        default:
          $ref: "#/components/responses/Error"
  /runs/{id}/events:
    get:
      tags: [events]
      operationId: listRunEvents
      summary: REST-доступ к журналу событий рана
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: after_id
          in: query
          schema:
            type: integer
            format: int64
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: страница событий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        default:
          $ref: "#/components/responses/Error"
  /runs/{id}/artifacts/{artifactId}/content:
    get:
      tags: [runs]
      operationId: getArtifactContent
      summary: |
        Содержимое файла артефакта (F-02, fix-task-4). Путь читается только
        из записи БД и обязан оставаться внутри каталога рана (run_dir);
        чужой/несуществующий артефакт, выход за run_dir и отсутствующий файл
        → 404. Файл больше cap'а (5 МБ) → 413 (фронт предлагает скачать).
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: artifactId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: содержимое файла (UTF-8)
          content:
            text/plain:
              schema:
                type: string
        "413":
          description: файл больше cap'а (5 МБ)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /gates/{id}/resolve:
    post:
      tags: [gates]
      operationId: resolveGate
      summary: Резолв гейта (идемпотентно; повтор → 200 + already_resolved)
      parameters:
        - $ref: "#/components/parameters/GateId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveGateRequest"
      responses:
        "200":
          description: гейт резолвнут
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveGateResponse"
        default:
          $ref: "#/components/responses/Error"
  /stages/{id}/interrupt:
    post:
      tags: [runs]
      operationId: interruptStage
      summary: Interrupt & Steer (D-22) — прервать этап и резюмить с сообщением
      parameters:
        - $ref: "#/components/parameters/StageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterruptStageRequest"
      responses:
        "200":
          description: этап прерван, steer-заметка поставлена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stage"
        default:
          $ref: "#/components/responses/Error"
  /telegram/pair:
    post:
      tags: [telegram]
      operationId: pairTelegram
      summary: Привязка TG-чата по коду из /start (T-19)
      description: |
        Пользователь отправляет боту /start → бот показывает 6-значный код →
        код вводится в UI → chat_id попадает в whitelist (tg_chats).
        Код одноразовый: успешная привязка его погашает.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairTelegramRequest"
      responses:
        "200":
          description: чат привязан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairTelegramResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /fs/browse:
    get:
      tags: [system]
      operationId: fsBrowse
      summary: Листинг подкаталогов (выбор папки проекта в браузерном UI, T-16)
      parameters:
        - name: path
          in: query
          schema:
            type: string
          description: абсолютный путь; пусто = домашний каталог
      responses:
        "200":
          description: подкаталоги
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FsBrowseResult"
        default:
          $ref: "#/components/responses/Error"
...
  /ws:
    get:
      tags: [events]
      operationId: eventsWs
      summary: WebSocket-поток событий журнала (D-04)
      description: |
        Upgrade в WebSocket. Протокол: догон событиями с id > last_event_id,
        сообщение `synced` (граница replay/live), затем live-поток Event'ов.
        Медленный клиент отключается (close 1001, "resync required") и
        переподключается с последним полученным last_event_id.
      parameters:
        - name: run_id
          in: query
          schema:
            type: string
            default: "*"
          description: ран для подписки; "*" — все раны (инбокс гейтов)
        - name: last_event_id
          in: query
          schema:
            type: integer
            format: int64
            default: 0
        - name: token
          in: query
          schema:
            type: string
          description: токен localhost-API (D-08), если настроен
      responses:
        "101":
          description: WebSocket upgrade; далее Event | SyncedMessage
        "401":
          description: неверный/отсутствующий token
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    PipelineId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Error:
      description: ошибка в едином формате
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Project:
      type: object
      required: [id, path, name, default_branch, ide_command, created_at]
      properties:
        id:
          type: integer
          format: int64
        path:
          type: string
        name:
          type: string
        default_branch:
          type: string
        ide_command:
          type: string
        notify_tg_default:
          type: boolean
          default: true
          description: дефолт notify_tg для новых ранов проекта (T-16)
        created_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [path, name]
      properties:
        path:
          type: string
        name:
          type: string
        default_branch:
          type: string
        ide_command:
          type: string
    PatchProjectRequest:
      type: object
      properties:
        default_branch:
          type: string
        ide_command:
          type: string
        notify_tg_default:
          type: boolean
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          required: [active_runs, open_gates]
          properties:
            active_runs:
              type: array
              items:
                $ref: "#/components/schemas/Run"
            open_gates:
              type: integer
              description: открытые гейты по всем активным ранам проекта
...
    RunState:
      type: string
      enum: [draft, running, waiting_gate, succeeded, failed, stopped]
...